Separate enter and exit events.
Currently, if a rule refers to evt.type=open, both enter and exit events
are passed to the rule, even if the rule also explicitly specifies a
direction using evt.dir.

Make it possible to separately handle enter vs exit events by keeping
separate tables of enter vs exit events. If a rule doesn't mention a
direction, or mentions both directions, both sets of events are
associated with the rule. However, if the rule does mention a direction,
only that direction's event ids are associated with the rule.